requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
paho-mqtt
python-dotenv
requests
pydantic-settings
paho-mqtt
